+++ packages/apptive_grid_form/test/apptive_grid_form_test.dart
@@ -33,7 +33,7 @@ void main() {
 
   setUp(() {
     client = MockApptiveGridClient();
-    when(() => client.sendPendingActions()).thenAnswer((_) async {});
+    when(() => client.sendPendingActions()).thenAnswer((_) async => []);
   });
 
   group('Title', () {
@@ -870,6 +870,109 @@ void main() {
       );
     });
 
+    testWidgets('Cache, Callback specified and return true, Shows Saved Screen',
+        (tester) async {
+      final callbackCompleter = Completer<ApptiveLink>();
+      final cacheMap = <ActionItem>{};
+      final cache = MockApptiveGridCache();
+      when(() => cache.addPendingActionItem(any())).thenAnswer(
+        (invocation) => cacheMap.add(invocation.positionalArguments[0]),
+      );
+      when(() => cache.removePendingActionItem(any())).thenAnswer(
+        (invocation) => cacheMap.remove(invocation.positionalArguments[0]),
+      );
+      when(() => cache.getPendingActionItems())
+          .thenAnswer((invocation) => cacheMap.toList());
+
+      final client = ApptiveGridClient(
+        httpClient: httpClient,
+        options: ApptiveGridOptions(
+          cache: cache,
+        ),
+      );
+
+      final target = TestApp(
+        client: client,
+        child: ApptiveGridForm(
+          uri: formUri,
+          onSavedToPending: (link, __) async {
+            callbackCompleter.complete(link);
+            return true;
+          },
+        ),
+      );
+
+      await tester.pumpWidget(target);
+      await tester.pumpAndSettle();
+
+      await tester.tap(find.byType(ActionButton));
+      await tester.pumpAndSettle();
+
+      verify(() => httpClient.send(any())).called(1);
+
+      expect(
+        find.text(
+          'The form was saved and will be sent at the next opportunity',
+          skipOffstage: false,
+        ),
+        findsOneWidget,
+      );
+
+      expect(await callbackCompleter.future, cacheMap.first.link);
+    });
+
+    testWidgets(
+        'Cache, Callback specified and return false, Does not show Saved Screen',
+        (tester) async {
+      final callbackCompleter = Completer<ApptiveLink>();
+      final cacheMap = <ActionItem>{};
+      final cache = MockApptiveGridCache();
+      when(() => cache.addPendingActionItem(any())).thenAnswer(
+        (invocation) => cacheMap.add(invocation.positionalArguments[0]),
+      );
+      when(() => cache.removePendingActionItem(any())).thenAnswer(
+        (invocation) => cacheMap.remove(invocation.positionalArguments[0]),
+      );
+      when(() => cache.getPendingActionItems())
+          .thenAnswer((invocation) => cacheMap.toList());
+
+      final client = ApptiveGridClient(
+        httpClient: httpClient,
+        options: ApptiveGridOptions(
+          cache: cache,
+        ),
+      );
+
+      final target = TestApp(
+        client: client,
+        child: ApptiveGridForm(
+          uri: formUri,
+          onSavedToPending: (link, __) async {
+            callbackCompleter.complete(link);
+            return false;
+          },
+        ),
+      );
+
+      await tester.pumpWidget(target);
+      await tester.pumpAndSettle();
+
+      await tester.tap(find.byType(ActionButton));
+      await tester.pumpAndSettle();
+
+      verify(() => httpClient.send(any())).called(1);
+
+      expect(
+        find.text(
+          'The form was saved and will be sent at the next opportunity',
+          skipOffstage: false,
+        ),
+        findsNothing,
+      );
+
+      expect(await callbackCompleter.future, cacheMap.first.link);
+    });
+
     testWidgets('Cache, Error in Attachment, Shows Saved Screen',
         (tester) async {
       final cacheMap = <ActionItem>{};
@@ -887,7 +990,8 @@ void main() {
       when(() => client.loadForm(uri: formUri)).thenAnswer(
         (invocation) => Future.value(data),
       );
-      when(() => client.sendPendingActions()).thenAnswer((invocation) async {});
+      when(() => client.sendPendingActions())
+          .thenAnswer((invocation) async => []);
       when(() => client.submitFormWithProgress(action, data)).thenAnswer(
         (invocation) => Stream.value(
           AttachmentCompleteProgressEvent(http.Response('', 400)),
@@ -1231,7 +1335,7 @@ void main() {
       final globalKey = GlobalKey<_ChangingFormWidgetState>();
       final client = MockApptiveGridClient();
 
-      when(client.sendPendingActions).thenAnswer((_) async {});
+      when(client.sendPendingActions).thenAnswer((_) async => []);
       when(() => client.loadForm(uri: any(named: 'uri'))).thenAnswer(
         (_) async => FormData(
           id: 'formId',