+++ packages/apptive_grid_form/test/enum_collection_form_widget_test.dart
@@ -107,7 +107,7 @@ void main() {
         );
         final client = MockApptiveGridClient();
         when(() => client.sendPendingActions())
-            .thenAnswer((_) => Future.value());
+            .thenAnswer((_) => Future.value([]));
         when(() => client.submitFormWithProgress(action, any())).thenAnswer(
           (_) => Stream.value(SubmitCompleteProgressEvent(Response('', 200))),
         );
@@ -158,7 +158,7 @@ void main() {
         );
         final client = MockApptiveGridClient();
         when(() => client.sendPendingActions())
-            .thenAnswer((_) => Future.value());
+            .thenAnswer((_) => Future.value([]));
         when(() => client.submitFormWithProgress(action, any())).thenAnswer(
           (_) => Stream.value(SubmitCompleteProgressEvent(Response('', 200))),
         );
@@ -301,7 +301,7 @@ void main() {
         );
         final client = MockApptiveGridClient();
         when(() => client.sendPendingActions())
-            .thenAnswer((_) => Future.value());
+            .thenAnswer((_) => Future.value([]));
         when(() => client.submitFormWithProgress(action, any())).thenAnswer(
           (_) => Stream.value(SubmitCompleteProgressEvent(Response('', 200))),
         );
@@ -353,7 +353,7 @@ void main() {
         );
         final client = MockApptiveGridClient();
         when(() => client.sendPendingActions())
-            .thenAnswer((_) => Future.value());
+            .thenAnswer((_) => Future.value([]));
         when(() => client.submitFormWithProgress(action, any())).thenAnswer(
           (_) => Stream.value(SubmitCompleteProgressEvent(Response('', 200))),
         );