+++ packages/apptive_grid_form/test/enum_form_widget_test.dart
@@ -43,7 +43,7 @@ void main() {
         );
         final client = MockApptiveGridClient();
         when(() => client.sendPendingActions())
-            .thenAnswer((_) => Future.value());
+            .thenAnswer((_) => Future.value([]));
         when(() => client.submitFormWithProgress(action, any())).thenAnswer(
           (_) =>
               Stream.value(SubmitCompleteProgressEvent(Response('body', 200))),
@@ -94,7 +94,7 @@ void main() {
         );
         final client = MockApptiveGridClient();
         when(() => client.sendPendingActions())
-            .thenAnswer((_) => Future.value());
+            .thenAnswer((_) => Future.value([]));
         when(() => client.submitFormWithProgress(action, any())).thenAnswer(
           (_) =>
               Stream.value(SubmitCompleteProgressEvent(Response('body', 200))),
@@ -141,7 +141,7 @@ void main() {
         );
         final client = MockApptiveGridClient();
         when(() => client.sendPendingActions())
-            .thenAnswer((_) => Future.value());
+            .thenAnswer((_) => Future.value([]));
         when(() => client.submitFormWithProgress(action, any())).thenAnswer(
           (_) =>
               Stream.value(SubmitCompleteProgressEvent(Response('body', 200))),
@@ -188,7 +188,7 @@ void main() {
         );
         final client = MockApptiveGridClient();
         when(() => client.sendPendingActions())
-            .thenAnswer((_) => Future.value());
+            .thenAnswer((_) => Future.value([]));
         when(() => client.submitFormWithProgress(action, any())).thenAnswer(
           (_) =>
               Stream.value(SubmitCompleteProgressEvent(Response('body', 200))),
@@ -243,7 +243,7 @@ void main() {
         );
         final client = MockApptiveGridClient();
         when(() => client.sendPendingActions())
-            .thenAnswer((_) => Future.value());
+            .thenAnswer((_) => Future.value([]));
         when(() => client.submitFormWithProgress(action, any())).thenAnswer(
           (_) =>
               Stream.value(SubmitCompleteProgressEvent(Response('body', 200))),