+++ packages/apptive_grid_form/test/phone_number_form_widget_test.dart
@@ -50,7 +50,7 @@ void main() {
         fields: [],
       ),
     );
-    when(() => client.sendPendingActions()).thenAnswer((_) => Future.value());
+    when(() => client.sendPendingActions()).thenAnswer((_) => Future.value([]));
     when(() => client.submitFormWithProgress(action, any())).thenAnswer(
       (_) => Stream.value(SubmitCompleteProgressEvent(Response('body', 200))),
     );