+++ packages/apptive_grid_user_management/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.4.0
+* Update apptive_grid_core to 0.15.0
+
 ## 0.3.1
 * Upgrade apptive_grid_core to 0.14.4
 * Fix Lints for Flutter 3.7.0