+++ packages/apptive_grid_web_apptive/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.4.0
+* Update apptive_grid_core to 0.15.0
+
 ## 0.3.1+1
 * Use `const` where possible